reporter: Move the evaluated model classes into separate files
To improve overview, because the single file became too large.
