+++ reporter/src/main/kotlin/model/EvaluatedModel.kt
@@ -21,8 +21,6 @@ package com.here.ort.reporter.model
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo
 import com.fasterxml.jackson.annotation.JsonIdentityReference
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.SerializationFeature
@@ -35,19 +33,10 @@ import com.fasterxml.jackson.dataformat.yaml.YAMLMapper
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 
 import com.here.ort.model.CustomData
-import com.here.ort.model.Identifier
-import com.here.ort.model.LicenseSource
 import com.here.ort.model.OrtIssue
 import com.here.ort.model.OrtResult
 import com.here.ort.model.PROPERTY_NAMING_STRATEGY
-import com.here.ort.model.PackageCurationResult
-import com.here.ort.model.PackageLinkage
-import com.here.ort.model.Provenance
-import com.here.ort.model.RemoteArtifact
 import com.here.ort.model.RuleViolation
-import com.here.ort.model.ScannerDetails
-import com.here.ort.model.Severity
-import com.here.ort.model.VcsInfo
 import com.here.ort.model.config.IssueResolution
 import com.here.ort.model.config.LicenseFindingCuration
 import com.here.ort.model.config.PathExclude
@@ -58,12 +47,8 @@ import com.here.ort.reporter.Reporter
 import com.here.ort.reporter.ReporterInput
 import com.here.ort.reporter.reporters.WebAppReporter
 import com.here.ort.reporter.utils.IntIdModule
-import com.here.ort.reporter.utils.ZeroBasedIntSequenceGenerator
-import com.here.ort.spdx.SpdxExpression
 
 import java.io.Writer
-import java.time.Instant
-import java.util.SortedSet
 
 /**
  * The [EvaluatedModel] represents the outcome of the evaluation of a [ReporterInput]. This means that all additional
@@ -182,147 +167,3 @@ data class EvaluatedModel(
         return tree
     }
 }
-
-data class EvaluatedPackage(
-    val id: Identifier,
-    val isProject: Boolean,
-    val definitionFilePath: String,
-    val purl: String = id.toPurl(),
-    val declaredLicenses: List<LicenseId>,
-    val declaredLicensesProcessed: EvaluatedProcessedDeclaredLicense,
-    val detectedLicenses: Set<LicenseId>,
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val concludedLicense: SpdxExpression? = null,
-    val description: String,
-    val homepageUrl: String,
-    val binaryArtifact: RemoteArtifact,
-    val sourceArtifact: RemoteArtifact,
-    val vcs: VcsInfo,
-    val vcsProcessed: VcsInfo = vcs.normalize(),
-    val curations: List<PackageCurationResult>,
-    @JsonIdentityReference(alwaysAsId = true)
-    val paths: MutableList<EvaluatedPackagePath>,
-    val levels: SortedSet<Int>,
-    val scopes: MutableSet<ScopeName>,
-    val scanResults: List<EvaluatedScanResult>,
-    val findings: List<EvaluatedFinding>,
-    val isExcluded: Boolean,
-    val pathExcludes: List<PathExclude>,
-    val scopeExcludes: List<ScopeExclude>,
-    val issues: List<EvaluatedOrtIssue>
-)
-
-data class EvaluatedPackagePath(
-    val pkg: EvaluatedPackage,
-    val project: EvaluatedPackage,
-    val scope: ScopeName,
-    val path: List<EvaluatedPackage>
-)
-
-data class EvaluatedScanResult(
-    val provenance: Provenance,
-    val scanner: ScannerDetails,
-    val startTime: Instant,
-    val endTime: Instant,
-    val fileCount: Int,
-    val packageVerificationCode: String,
-    val issues: List<EvaluatedOrtIssue>
-)
-
-enum class EvaluatedFindingType {
-    COPYRIGHT, LICENSE
-}
-
-data class EvaluatedFinding(
-    val type: EvaluatedFindingType,
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val license: LicenseId?,
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val copyright: CopyrightStatement?,
-    val path: String,
-    val startLine: Int,
-    val endLine: Int,
-    val scanResult: EvaluatedScanResult
-)
-
-/**
- * Wrapper class for copyright statements. Allows Jackson to generate IDs for them when storing them in a separate list
- * for de-duplication.
- */
-data class CopyrightStatement(
-    val statement: String
-)
-
-/**
- * Wrapper class for license identifiers. Allows Jackson to generate IDs for them when storing them in a separate list
- * for de-duplication.
- */
-data class LicenseId(
-    val id: String
-)
-
-/**
- * Wrapper class for scope names. Allows Jackson to generate IDs for them when storing them in a separate list for
- * de-duplication.
- */
-data class ScopeName(
-    val name: String
-)
-
-data class EvaluatedProcessedDeclaredLicense(
-    val spdxExpression: SpdxExpression?,
-    val mappedLicenses: List<LicenseId>,
-    val unmappedLicenses: List<LicenseId>
-)
-
-enum class EvaluatedOrtIssueType {
-    ANALYZER, SCANNER
-}
-
-@JsonIgnoreProperties(value = ["pkg", "scanResult", "path"], allowGetters = true)
-data class EvaluatedOrtIssue(
-    val timestamp: Instant,
-    val type: EvaluatedOrtIssueType,
-    val source: String,
-    val message: String,
-    val severity: Severity = Severity.ERROR,
-    val resolutions: List<IssueResolution>,
-    @JsonIdentityReference(alwaysAsId = true)
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val pkg: EvaluatedPackage?,
-    @JsonIdentityReference(alwaysAsId = true)
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val scanResult: EvaluatedScanResult?,
-    @JsonIdentityReference(alwaysAsId = true)
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val path: EvaluatedPackagePath?
-)
-
-data class EvaluatedRuleViolation(
-    val rule: String,
-    val pkg: EvaluatedPackage,
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val license: LicenseId?,
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val licenseSource: LicenseSource?,
-    val severity: Severity,
-    val message: String,
-    val howToFix: String,
-    val resolutions: List<RuleViolationResolution>
-)
-
-@JsonIdentityInfo(property = "key", generator = ZeroBasedIntSequenceGenerator::class, scope = DependencyTreeNode::class)
-data class DependencyTreeNode(
-    val title: String,
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val linkage: PackageLinkage?,
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    val pkg: EvaluatedPackage?,
-    @JsonInclude(JsonInclude.Include.NON_EMPTY)
-    val pathExcludes: List<PathExclude> = emptyList(),
-    @JsonInclude(JsonInclude.Include.NON_EMPTY)
-    val scopeExcludes: List<ScopeExclude> = emptyList(),
-    @JsonInclude(JsonInclude.Include.NON_EMPTY)
-    val issues: List<EvaluatedOrtIssue> = emptyList(),
-    val children: List<DependencyTreeNode>
-)
