+++ reporter/src/main/kotlin/model/EvaluatedProcessedDeclaredLicense.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright (C) 2017-2020 HERE Europe B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package com.here.ort.reporter.model
+
+import com.here.ort.spdx.SpdxExpression
+
+data class EvaluatedProcessedDeclaredLicense(
+    val spdxExpression: SpdxExpression?,
+    val mappedLicenses: List<LicenseId>,
+    val unmappedLicenses: List<LicenseId>
+)
